- implemented basic logic

+++ employee-service/src/main/java/com/mlavrenko/api/service/EmployeeService.java
@@ -4,19 +4,24 @@ import com.mlavrenko.api.domain.Department;
 import com.mlavrenko.api.domain.Employee;
 import com.mlavrenko.api.dto.DepartmentDTO;
 import com.mlavrenko.api.dto.EmployeeDTO;
-import com.mlavrenko.api.notification.Message;
+import com.mlavrenko.api.notification.EmployeeEvent;
 import com.mlavrenko.api.notification.EmployeeStatus;
 import com.mlavrenko.api.repository.EmployeeRepository;
+
 import java.util.UUID;
 import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 
 @Service
 @Transactional
 public class EmployeeService {
+    private final Logger logger = LoggerFactory.getLogger(getClass());
     private final EmployeeRepository employeeRepository;
     private final DepartmentService departmentService;
     private final MessageService messageService;
@@ -28,17 +33,19 @@ public class EmployeeService {
     }
 
     public EmployeeDTO createEmployee(@Valid EmployeeDTO employeeDTO) {
+        logger.info("Create new employee {}", employeeDTO);
         Employee employee = convertToDomain(employeeDTO);
         Employee saved = employeeRepository.save(employee);
-        messageService.sendMessage(new Message(saved.getUuid(), EmployeeStatus.CREATED));
+        messageService.sendMessage(new EmployeeEvent(saved.getId(), EmployeeStatus.CREATED));
         return convertToDto(saved);
     }
 
     public EmployeeDTO updateEmployee(@Valid EmployeeDTO employeeDTO) {
-        Employee employee = getById(employeeDTO.getUuid());
+        logger.info("Update employee {}", employeeDTO);
+        Employee employee = getById(employeeDTO.getId());
         copyToEntity(employeeDTO, employee);
         Employee saved = employeeRepository.save(employee);
-        messageService.sendMessage(new Message(saved.getUuid(), EmployeeStatus.UPDATED));
+        messageService.sendMessage(new EmployeeEvent(saved.getId(), EmployeeStatus.UPDATED));
         return convertToDto(saved);
     }
 
@@ -47,12 +54,21 @@ public class EmployeeService {
     }
 
     public void deleteEmployee(@Valid UUID uuid) {
+        logger.info("Delete employee by id={}", uuid);
+        if (!employeeRepository.existsById(uuid)) {
+            throw new EntityNotFoundException();
+        }
         employeeRepository.deleteById(uuid);
-        messageService.sendMessage(new Message(uuid, EmployeeStatus.DELETED));
+        messageService.sendMessage(new EmployeeEvent(uuid, EmployeeStatus.DELETED));
     }
 
     public EmployeeDTO getEmployeeById(@Valid UUID uuid) {
-        return convertToDto(getById(uuid));
+        logger.info("Get employee by id={}", uuid);
+        Employee byId = getById(uuid);
+        logger.info("Retrived entity={}", byId);
+        EmployeeDTO employeeDTO = convertToDto(byId);
+        logger.info("Retrived dto={}", employeeDTO);
+        return employeeDTO;
     }
 
     private void copyToEntity(@Valid EmployeeDTO employeeDTO, Employee employee) {
@@ -62,7 +78,7 @@ public class EmployeeService {
 
     private Employee convertToDomain(EmployeeDTO employeeDTO) {
         Employee employee = new Employee();
-        BeanUtils.copyProperties(employeeDTO, employee, "uuid", "department");
+        BeanUtils.copyProperties(employeeDTO, employee, "department");
         Department department = departmentService.getById(employeeDTO.getDepartment().getId()).orElseThrow(IllegalArgumentException::new);
         employee.setDepartment(department);
         return employee;
